Renamed UserFormDTO to UserDataDTO to better represent the class purpose. Moved the class into a new package.
