+++ src/main/java/org/bitmaxsystems/oop2library/util/contracts/IUserFormChain.java
@@ -1,10 +1,10 @@
 package org.bitmaxsystems.oop2library.util.contracts;
 
-import org.bitmaxsystems.oop2library.models.form.UserFormDTO;
+import org.bitmaxsystems.oop2library.models.dto.UserDataDTO;
 
 public interface IUserFormChain {
 
     void setNextChain (IUserFormChain chain);
 
-    void execute(UserFormDTO formData) throws Exception;
+    void execute(UserDataDTO formData) throws Exception;
 }
