+++ src/main/java/org/bitmaxsystems/oop2library/util/userformchain/UpdatePasswordChain.java
@@ -1,7 +1,7 @@
 package org.bitmaxsystems.oop2library.util.userformchain;
 
 import org.bitmaxsystems.oop2library.models.auth.Credentials;
-import org.bitmaxsystems.oop2library.models.form.UserFormDTO;
+import org.bitmaxsystems.oop2library.models.dto.UserDataDTO;
 import org.bitmaxsystems.oop2library.repository.GenericRepository;
 import org.bitmaxsystems.oop2library.util.contracts.IUserFormChain;
 import org.springframework.security.crypto.bcrypt.BCrypt;
@@ -16,7 +16,7 @@ public class UpdatePasswordChain implements IUserFormChain {
     }
 
     @Override
-    public void execute(UserFormDTO formData) throws Exception {
+    public void execute(UserDataDTO formData) throws Exception {
         if (formData.isNewPassword())
         {
             Credentials credentials = formData.getUser().getCredentials();
